Working Version 0.1
This version works for CH4 in STO-3G and 6-31G bases, and for C2H2 in STO-3G basis. Probably for all molecules containing just the atoms defined in the Build_Basis function. The Build_Basis_2 function is not yet finished, but is not currently used. If something doesn't work with it just comment that line.
The same with lines involving the output of Build_ERI_fast. There is a plot involving it which should be removed